+++ packages/windie-sdk-js/cjs/transport/backendPayloadContract.js
@@ -37,6 +37,7 @@ const BACKEND_PAYLOAD_KEYS_BY_TYPE = Object.freeze({
         'agent_full_sudo_enabled',
         'browser_automation_enabled',
         'include_query_screenshot',
+        'tools',
         'provider_api_keys',
         'provider_oauth',
     ]),